REFACTOR `Autocomplete` & `Extract` into Services namespace
